Update web.py
Added validation for installation of volttron-ansible packages. Fixed validation for ansible and git packages.

+++ web.py
@@ -11,22 +11,17 @@ if sys.version_info < (3, 10):
     print("The lastest version of Python 3 has been installed.")
 
 # Check if the ansible and git package are installed; Install ansible and git if not installed
-print("Now checking if the ansible package is installed.")
+print("Now checking if the ansible and git packages are installed.")
 dpkg_output = subprocess.check_output(['dpkg', '-l'])
 
 packages = [line.split()[1] for line in dpkg_output.decode().splitlines() if line.startswith('ii')]
+ansible_installed = False
+git_installed = False
 for package in packages:
     if package == "ansible":
         ansible_installed = True
-        break
-    else:
-        ansible_installed = False
-
     if package == "git":
         git_installed = True
-        break
-    else:
-        git_installed =False
     
 if not ansible_installed:
     print("Ansible has not been installed. Now installing the ansible package using apt.")
@@ -48,7 +43,7 @@ elif git_installed:
     If there is a virtual environment under .venv, ask the user if they want to 
     override the current .venv or name the virtual environment as something else.
 '''
-print("Now checking if the virtual environment '.venv' exists")
+print("Now checking if the virtual environment '.venv' exists.")
 if not os.path.exists('.venv'):
     print("The virtual environment '.venv' does not exist. Now creating the virtual environment as .venv")
     Popen(['bash', '-c', 'python3 -m venv .venv']).wait()
@@ -76,10 +71,20 @@ elif os.path.exists('.venv'):
 
 # Activate the virtual environment and install volttron-ansible depending on whether the venv was overriden or not.
 if override:
-    print("Now activating the virtual environment '.venv' and installing the volttron-ansible package.")
-    Popen(['bash', '-c', 'source .venv/bin/activate && ansible-galaxy install git+https://github.com/volttron/volttron-ansible.git']).wait()
-    print("The volttron-ansible package has been installed inside the virtual environment '.venv.'")
-if not override:
-    print(f"Now activating the virtual environemnt '{venv_name}' and installing the volttron-ansible package.")
-    Popen(['bash', '-c', f'source {venv_name}/bin/activate && ansible-galaxy install git+https://github.com/volttron/volttron-ansible.git']).wait()
-    print(f"The volttron-ansible package has been installed inside the virtual environment '{venv_name}'.")
+    print("Now checking if the package 'volttron-ansible' is installed.")
+    if not os.path.exists(os.path.expanduser("~") + "/.ansible/roles/volttron-ansible"):
+        print("The package 'volttron-ansible' is not installed.")
+        print("Now activating the virtual environment '.venv' and installing the package 'volttron-ansible'.")
+        Popen(['bash', '-c', 'source .venv/bin/activate && ansible-galaxy install git+https://github.com/volttron/volttron-ansible.git']).wait()
+        print("The package 'volttron-ansible' has been installed inside the virtual environment '.venv'.")
+    else:
+        print("The package 'volttron-ansible' is already installed.")
+else:
+    print("Now checking if the package 'volttron-ansible' is installed.")
+    if not os.path.exists(os.path.expanduser("~") + "/.ansible/roles/volttron-ansible"):
+        print("The package 'volttron-ansible' has not been installed.")
+        print(f"Now activating the virtual environemnt '{venv_name}' and installing the package 'volttron-ansible'.")
+        Popen(['bash', '-c', f'source {venv_name}/bin/activate && ansible-galaxy install git+https://github.com/volttron/volttron-ansible.git']).wait()
+        print(f"The package 'volttron-ansible' has been installed inside the virtual environment '{venv_name}'.")
+    else:
+        print("The package 'volttron-ansible' is already installed.")
